add transform util func to async-node-plugin
