+++ plugins/async-node/core/BUILD
@@ -14,10 +14,15 @@ js_pipeline(
     native_bundle = "AsyncNodePlugin",
     peer_deps = [
         ":node_modules/@player-ui/player",
+        ":node_modules/@player-ui/react",
+        "//:node_modules/react",
+        "//:node_modules/@types/react",
     ],
     deps = [
         "//:node_modules/queue-microtask",
         "//:node_modules/tapable-ts",
         "//:node_modules/timm",
+        "//:node_modules/uuid",
+        "//:node_modules/@types/uuid",
     ],
 )
