+++ plugins/async-node/core/package.json
@@ -3,7 +3,14 @@
     "version": "0.0.0-PLACEHOLDER",
     "main": "src/index.ts",
     "peerDependencies": {
-      "@player-ui/player": "workspace:*"
+      "@player-ui/player": "workspace:*",
+      "@player-ui/react": "workspace:*"
+    },
+    "dependencies": {
+      "uuid": "^8.3.2"
+    },
+    "devDependencies": {
+      "@types/uuid": "^8.3.4"
     }
   }
