+++ plugins/reference-assets/core/src/assets/chat-message/transform.ts
@@ -2,11 +2,9 @@ import type {
   BeforeTransformFunction,
   TransformFunctions,
 } from "@player-ui/player";
-import { v4 as uuid } from "uuid";
-import { Builder } from "@player-ui/player";
 import { composeBefore, compose } from "@player-ui/asset-transform-plugin";
 import type { ChatMessageAsset } from "./types";
-
+import { asyncTransform } from "@player-ui/async-node-plugin";
 /**
  * In beforeTransform function, pass in flatten marker and call beforeResolve function
  * input: asset
@@ -34,23 +32,7 @@ import type { ChatMessageAsset } from "./types";
  */
 
 export const transform: BeforeTransformFunction<ChatMessageAsset> = (asset) => {
-  const id = uuid();
-
-  const assetNode = Builder.asset({ ...asset.value, type: "text" });
-  const asyncNode = Builder.asyncNode(id);
-  const multiNode = Builder.multiNode(
-    asset.value.flatten,
-    assetNode,
-    asyncNode,
-  );
-
-  const wrapperAsset = Builder.asset({
-    id: "chat-message-wrapper",
-    type: "chat-message-wrapper",
-  });
-
-  Builder.addChild(wrapperAsset, ["values"], multiNode);
-  return wrapperAsset;
+  return asyncTransform(asset, "text", "chat-message-wrapper");
 };
 
 export const chatMessageTransform: TransformFunctions = compose(
